+++ lib/controllers/rings_controller.dart
@@ -56,7 +56,7 @@ class RingsController extends GetxController {
     receiver.statusMenuItems = List.generate(34, (index) {
       return StatusMenuItem(
         name: Constants.names[index % 5],
-        time: Random().nextInt(3000).toDouble(),
+        time: Random().nextInt(3000),
       );
     });
     super.onInit();
